requirements.txt
# Document processing
langchain>=0.1.0
langchain-community>=0.0.10
beautifulsoup4>=4.12.0

# Embedding & Vector Store
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0

# LLM (optional - for generator step)
openai>=1.0.0

# Testing & Auto-grading
pytest>=7.4.0
pytest-cov>=4.1.0
